+++ assets/stylesheets/_components.scss
@@ -107,7 +107,7 @@
 @import 'components/segmented-control/style';
 @import 'components/select-dropdown/style';
 @import 'components/seo/style';
-@import 'my-sites/upgrade-nudge/expanded/style';
+@import 'blocks/upgrade-nudge-expanded/style';
 @import 'components/seo/preview-upgrade-nudge/style';
 @import 'components/signup-form/style';
 @import 'components/site-icon/style';
