+++ client/blocks/upgrade-nudge-expanded/style.scss
