+++ client/my-sites/upgrade-nudge/docs/example.jsx
@@ -7,10 +7,6 @@ import React from 'react';
  * Internal dependencies
  */
 import UpgradeNudge from 'my-sites/upgrade-nudge';
-import ExpandedUpgradeNudge from 'my-sites/upgrade-nudge/expanded';
-import {
-	PLAN_BUSINESS, FEATURE_ADVANCED_SEO
-} from 'lib/plans/constants';
 
 export default React.createClass( {
 
@@ -34,21 +30,6 @@ export default React.createClass( {
 						message="This is a custom message"
 						icon="customize"
 						compact
-					/>
-				</div>
-				<div>
-					<ExpandedUpgradeNudge
-						plan={ PLAN_BUSINESS }
-						upgrade={ () => alert( 'Props for upgrading!' ) }
-						title={ this.translate( 'Title' ) }
-						subtitle={ this.translate( 'Subtitle' ) }
-						highlightedFeature={ FEATURE_ADVANCED_SEO }
-						event={ 'example' }
-						benefits={ [
-							'First benefit',
-							'Second benefit',
-							'Third benefit'
-						] }
 					/>
 				</div>
 			</div>
